+++ src/chargebee-resource.class.ts
@@ -1,9 +1,15 @@
-import { ChargeBee } from "chargebee-typescript";
+import type { ChargeBee } from "chargebee-typescript";
 
 import type { RequestWrapper } from "chargebee-typescript/lib/request_wrapper";
 import type { ListResult } from "chargebee-typescript/lib/list_result";
 import type { Result } from "chargebee-typescript/lib/result";
-import type { ProcessWait } from "chargebee-typescript/lib/process_wait";
+
+import type {
+  ResourceResult,
+  ResultMethodName,
+  ListResultMethodName,
+  ResolveResultReturn,
+} from "./chargebee-resource.types";
 
 export class ChargebeeResource {
   constructor(protected readonly chargebee: ChargeBee) {}
@@ -80,39 +86,3 @@ export class ChargebeeResource {
         Object.keys(returning).map((key) => [key, result[key as keyof Result]]),
       ) as ResolveResultReturn<TReturning>;
 }
-
-type ResultMethodName<
-  TResource extends keyof ChargeBee,
-  TMethod extends keyof ChargeBee[TResource],
-> = ChargeBee[TResource][TMethod] extends (
-  ...args: unknown[]
-) => RequestWrapper<infer R>
-  ? R extends Result | ProcessWait
-    ? TMethod
-    : "Method must return a RequestWrapper<Result | ProcessWait>"
-  : never;
-
-type ListResultMethodName<
-  TResource extends keyof ChargeBee,
-  TMethod extends keyof ChargeBee[TResource],
-> = ChargeBee[TResource][TMethod] extends (
-  ...args: unknown[]
-) => RequestWrapper<infer R>
-  ? R extends ListResult
-    ? TMethod
-    : "Method must return a RequestWrapper<ListResult>"
-  : never;
-
-type ResourceResult = {
-  [K in keyof Result]?: {
-    optional: boolean;
-  };
-};
-
-type ResolveResultReturn<T extends ResourceResult> = {
-  [K in keyof T]: K extends keyof Result
-    ? T[K] extends { optional: true }
-      ? Result[K] | undefined
-      : Result[K]
-    : never;
-};
