+++ src/resources/coupon-code-resource.ts
@@ -1,4 +1,4 @@
-import { ChargebeeResource } from "./abstract-resource";
+import { ChargebeeResource } from "../chargebee-resource.class";
 export class CouponCodeResource extends ChargebeeResource {
   public readonly create = super.request("coupon_code", "create", {
     coupon_code: { optional: false },
