+++ src/resources/coupon-resource.ts
@@ -1,4 +1,4 @@
-import { ChargebeeResource } from "./abstract-resource";
+import { ChargebeeResource } from "../chargebee-resource.class";
 export class CouponResource extends ChargebeeResource {
   public readonly create = super.request("coupon", "create", {
     coupon: { optional: false },
