+++ src/resources/customer-resource.ts
@@ -1,4 +1,4 @@
-import { ChargebeeResource } from "./abstract-resource";
+import { ChargebeeResource } from "../chargebee-resource.class";
 export class CustomerResource extends ChargebeeResource {
   public readonly create = super.request("customer", "create", {
     customer: { optional: false },
