+++ src/resources/estimate-resource.ts
@@ -1,4 +1,4 @@
-import { ChargebeeResource } from "./abstract-resource";
+import { ChargebeeResource } from "../chargebee-resource.class";
 export class EstimateResource extends ChargebeeResource {
   public readonly createSubscription = super.request(
     "estimate",
