+++ src/resources/event-resource.ts
@@ -1,4 +1,4 @@
-import { ChargebeeResource } from "./abstract-resource";
+import { ChargebeeResource } from "../chargebee-resource.class";
 export class EventResource extends ChargebeeResource {
   public readonly list = super.listRequest("event", "list", {
     event: { optional: false },
