+++ src/resources/invoice-resource.ts
@@ -1,4 +1,4 @@
-import { ChargebeeResource } from "./abstract-resource";
+import { ChargebeeResource } from "../chargebee-resource.class";
 export class InvoiceResource extends ChargebeeResource {
   public readonly create = super.request("invoice", "create", {
     invoice: { optional: false },
