+++ src/resources/item-resource.ts
@@ -1,4 +1,4 @@
-import { ChargebeeResource } from "./abstract-resource";
+import { ChargebeeResource } from "../chargebee-resource.class";
 export class ItemResource extends ChargebeeResource {
   public readonly create = super.request("item", "create", {
     item: { optional: false },
