+++ src/resources/non-subscription-resource.ts
@@ -1,4 +1,4 @@
-import { ChargebeeResource } from "./abstract-resource";
+import { ChargebeeResource } from "../chargebee-resource.class";
 export class NonSubscriptionResource extends ChargebeeResource {
   public readonly processReceipt = super.request(
     "non_subscription",
