+++ src/resources/order-resource.ts
@@ -1,4 +1,4 @@
-import { ChargebeeResource } from "./abstract-resource";
+import { ChargebeeResource } from "../chargebee-resource.class";
 export class OrderResource extends ChargebeeResource {
   public readonly create = super.request("order", "create", {
     order: { optional: false },
