+++ src/resources/payment-intent-resource.ts
@@ -1,4 +1,4 @@
-import { ChargebeeResource } from "./abstract-resource";
+import { ChargebeeResource } from "../chargebee-resource.class";
 export class PaymentIntentResource extends ChargebeeResource {
   public readonly create = super.request("payment_intent", "create", {
     payment_intent: { optional: false },
