+++ src/resources/plan-resource.ts
@@ -1,4 +1,4 @@
-import { ChargebeeResource } from "./abstract-resource";
+import { ChargebeeResource } from "../chargebee-resource.class";
 export class PlanResource extends ChargebeeResource {
   public readonly create = super.request("plan", "create", {
     plan: { optional: false },
