+++ src/resources/purchase-resource.ts
@@ -1,4 +1,4 @@
-import { ChargebeeResource } from "./abstract-resource";
+import { ChargebeeResource } from "../chargebee-resource.class";
 export class PurchaseResource extends ChargebeeResource {
   public readonly create = super.request("purchase", "create", {
     purchase: { optional: false },
