+++ src/resources/site-migration-detail-resource.ts
@@ -1,4 +1,4 @@
-import { ChargebeeResource } from "./abstract-resource";
+import { ChargebeeResource } from "../chargebee-resource.class";
 export class SiteMigrationDetailResource extends ChargebeeResource {
   public readonly list = super.listRequest("site_migration_detail", "list", {
     site_migration_detail: { optional: false },
