+++ src/resources/subscription-resource.ts
@@ -1,4 +1,4 @@
-import { ChargebeeResource } from "./abstract-resource";
+import { ChargebeeResource } from "../chargebee-resource.class";
 export class SubscriptionResource extends ChargebeeResource {
   public readonly create = super.request(
     "subscription",
