+++ src/resources/usage-resource.ts
@@ -1,4 +1,4 @@
-import { ChargebeeResource } from "./abstract-resource";
+import { ChargebeeResource } from "../chargebee-resource.class";
 export class UsageResource extends ChargebeeResource {
   public readonly create = super.request("usage", "create", {
     usage: { optional: false },
